requirements.txt
aiohttp>=3.8.4
aiohttp-retry==2.8.3
aioredis==2.0.1
aiosignal==1.3.1
amqp==5.2.0
asgiref>=3.5.0,<4
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer>=2.0,<4.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
djangochannelsrestframework==1.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.0
flower==2.0.1
frozenlist==1.4.0
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
humanize==4.9.0
Hypercorn==0.15.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
kombu==5.3.4
msgpack==1.0.7
multidict==6.0.4
packaging==23.2
Pillow==10.1.0
priority==2.0.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
taskgroup==0.0.0a4
tomli==2.0.1
tornado==6.4
twilio==8.10.3
Twisted==23.10.0
txaio==23.1.1
typing-extensions==4.8.0 
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
websocket-client==1.7.0
wsproto==1.2.0
yarl==1.9.4
zope.interface==6.1
